PP-10063: Remove redundant endpoints
* /v1/api/invites/otp/resend
* /v1/api/invites/{code}/otp/generate

These were only needed for the old registration journey.  All services which were only needed for these have also been removed.
