+++ .secrets.baseline
@@ -123,51 +123,51 @@
       {
         "type": "Secret Keyword",
         "filename": "openapi/adminusers_spec.yaml",
-        "hashed_secret": "0ea7458942ab65e0a340cf4fd28ca00d93c494f3",
-        "is_verified": false,
-        "line_number": 267
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "675fecb5af336276574d3d88a4d382d961cf7418",
         "is_verified": false,
-        "line_number": 335
+        "line_number": 285
       },
       {
         "type": "Secret Keyword",
         "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "fefda504251b73d30a7cf3d047d241388a3c11b3",
         "is_verified": false,
-        "line_number": 336
+        "line_number": 286
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "614770647df3ab100b871bfc0d20e72c8625a5c4",
         "is_verified": false,
-        "line_number": 448
+        "line_number": 398
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "1976a945a8d2733655e4b2453bd49fb59cb7ba19",
         "is_verified": false,
-        "line_number": 709
+        "line_number": 659
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "a31519136d26754afeb0df49b5f066f387091f88",
         "is_verified": false,
-        "line_number": 743
+        "line_number": 693
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "openapi/adminusers_spec.yaml",
+        "hashed_secret": "0ea7458942ab65e0a340cf4fd28ca00d93c494f3",
+        "is_verified": false,
+        "line_number": 792
       },
       {
         "type": "Hex High Entropy String",
         "filename": "openapi/adminusers_spec.yaml",
         "hashed_secret": "3660e32cde5fccc8d1e4521d0c831c2012388720",
         "is_verified": false,
-        "line_number": 1238
+        "line_number": 1188
       }
     ],
     "src/main/java/uk/gov/pay/adminusers/model/CreateUserRequest.java": [
@@ -251,7 +251,7 @@
         "filename": "src/main/java/uk/gov/pay/adminusers/resources/InviteRequestValidator.java",
         "hashed_secret": "5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8",
         "is_verified": false,
-        "line_number": 33
+        "line_number": 20
       }
     ],
     "src/main/java/uk/gov/pay/adminusers/resources/ResetPasswordResource.java": [
@@ -425,14 +425,14 @@
         "filename": "src/test/java/uk/gov/pay/adminusers/service/InviteServiceTest.java",
         "hashed_secret": "7300c52390b4026eb1a60642b0eabfb54475c21b",
         "is_verified": false,
-        "line_number": 60
+        "line_number": 59
       },
       {
         "type": "Secret Keyword",
         "filename": "src/test/java/uk/gov/pay/adminusers/service/InviteServiceTest.java",
         "hashed_secret": "6318553899daae2941718c02508aeee938af1a1c",
         "is_verified": false,
-        "line_number": 306
+        "line_number": 230
       }
     ],
     "src/test/java/uk/gov/pay/adminusers/service/ResetPasswordServiceTest.java": [
@@ -484,5 +484,5 @@
       }
     ]
   },
-  "generated_at": "2022-11-29T12:45:26Z"
+  "generated_at": "2022-12-08T15:27:16Z"
 }
