+++ src/test/java/uk/gov/pay/adminusers/resources/InviteResourceOtpIT.java
@@ -234,67 +234,4 @@ class InviteResourceOtpIT extends IntegrationTest {
                 .then()
                 .statusCode(PRECONDITION_FAILED.getStatusCode());
     }
-
-    @Test
-    void resendOtp_shouldUpdateTelephoneNumber_whenValidOtp() throws Exception {
-
-        // create an invitation with initial telephone number
-        String initialTelephoneNumber = "+447451111111";
-        code = InviteDbFixture.inviteDbFixture(databaseHelper)
-                .withEmail(EMAIL)
-                .withOtpKey(OTP_KEY)
-                .withTelephoneNumber(initialTelephoneNumber)
-                .withPassword(PASSWORD)
-                .expired()
-                .insertInviteToAddUserToService();
-
-        // generate new invitationOtpRequest with new telephone number
-        String newTelephoneNumber = "+447452222222";
-        Map<Object, Object> resendRequest = Map.of(
-                "code", code,
-                "telephone_number", newTelephoneNumber);
-
-        givenSetup()
-                .when()
-                .body(mapper.writeValueAsString(resendRequest))
-                .contentType(JSON)
-                .post(INVITES_RESEND_OTP_RESOURCE_URL)
-                .then()
-                .statusCode(NO_CONTENT.getStatusCode());
-
-        // check if we are using the newTelephoneNumber in the invitation
-        Map<String, Object> foundInvite = databaseHelper.findInviteByCode(code).get();
-        assertThat(foundInvite.get("telephone_number"), is(newTelephoneNumber));
-    }
-
-    @Test
-    void resendOtp_shouldFail_whenAllMandatoryFieldsAreMissing() throws Exception {
-
-        givenSetup()
-                .when()
-                .body(mapper.writeValueAsString(emptyMap()))
-                .contentType(JSON)
-                .post(INVITES_RESEND_OTP_RESOURCE_URL)
-                .then()
-                .statusCode(422)
-                .body("errors", hasSize(2))
-                .body("errors", hasItems("code must not be empty", "telephoneNumber must not be empty"));
-    }
-
-    @Test
-    void resendOtp_shouldFail_whenTelephoneNumberIsInvalid() throws Exception {
-        Map<Object, Object> resendRequest = Map.of(
-                "code", code,
-                "telephone_number", "invalid");
-
-        givenSetup()
-                .when()
-                .body(mapper.writeValueAsString(resendRequest))
-                .contentType(JSON)
-                .post(INVITES_RESEND_OTP_RESOURCE_URL)
-                .then()
-                .statusCode(422)
-                .body("errors", hasSize(1))
-                .body("errors", hasItems("telephoneNumber must be a valid telephone number"));
-    }
 }
